Rename power-grid test module and move helper inline
pytest's default import mode imports each test_*.py file as a
top-level module using its basename. packages/drinking-water
already uses test_model.py, so packages/power-grid/test_model.py
triggered ImportPathMismatchError during collection. Rename to
test_power_grid_model.py, matching the pattern set by
packages/transport-assignment/test_traffic_assignment_model.py.

get_dataset had been imported from conftest.py via a relative
import, which only works when tests/ is a package (__init__.py).
Removing that __init__.py in the previous commit left the import
broken. Since get_dataset is a plain helper used only in this
test file, move it alongside the tests; leave conftest.py with
fixtures only, matching the other package conftests.
